feat: rename package runner deployment
