+++ tdp/cli/commands/reconfigure.py
@@ -10,14 +10,14 @@ from tdp.cli.commands.queries import get_latest_success_service_component_versio
 from tdp.cli.session import get_session_class
 from tdp.cli.utils import check_services_cleanliness, collection_paths_to_collections
 from tdp.core.dag import Dag
-from tdp.core.models import StateEnum
-from tdp.core.runner import (
+from tdp.core.deployment import (
     AnsibleExecutor,
     DeploymentPlan,
     DeploymentRunner,
     EmptyDeploymentPlanError,
     NothingToRestartError,
 )
+from tdp.core.models import StateEnum
 from tdp.core.variables import ClusterVariables
