+++ tdp/cli/commands/resume.py
@@ -10,8 +10,8 @@ from tdp.cli.commands.queries import get_deployment, get_last_deployment
 from tdp.cli.session import get_session_class
 from tdp.cli.utils import check_services_cleanliness, collection_paths_to_collections
 from tdp.core.dag import Dag
+from tdp.core.deployment import AnsibleExecutor, DeploymentPlan, DeploymentRunner
 from tdp.core.models.state_enum import StateEnum
-from tdp.core.runner import AnsibleExecutor, DeploymentPlan, DeploymentRunner
 from tdp.core.variables import ClusterVariables
