+++ tdp/core/deployment/deployment_iterator.py
